[p2p] updated networking
